# supabase/schema-openapi.json
{
  "swagger": "2.0",
  "info": {
    "description": "",
    "title": "standard public schema",
    "version": "14.1"
  },
  "host": "wjcbobcqyqdkludsbqgf.supabase.co:443",
  "basePath": "/",
  "schemes": [
    "https"
  ],
  "consumes": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "produces": [
    "application/json",
    "application/vnd.pgrst.object+json;nulls=stripped",
    "application/vnd.pgrst.object+json",
    "text/csv"
  ],
  "paths": {
...
    "/cursos": {
      "get": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.cursos.id"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.descricaogeral"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.duracao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahoraria"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahorariaestagio"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahorariaatividadescomplementares"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.mediarequerida"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.frequenciarequerida"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.nivelensino"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.grauconferido"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.tituloconferido"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.valorinscricao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.valormensalidade"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.layoutnotas"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.idcenso"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.bibliotecavirtual"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.portariarecredenciamento"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.exibircrm"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.exibirbibliotecavirtual"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.situacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.numero_id"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/order"
          },
          {
            "$ref": "#/parameters/range"
          },
          {
            "$ref": "#/parameters/rangeUnit"
          },
          {
            "$ref": "#/parameters/offset"
          },
          {
            "$ref": "#/parameters/limit"
          },
          {
            "$ref": "#/parameters/preferCount"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "items": {
                "$ref": "#/definitions/cursos"
              },
              "type": "array"
            }
          },
          "206": {
            "description": "Partial Content"
          }
        },
        "tags": [
          "cursos"
        ]
      },
      "post": {
        "parameters": [
          {
            "$ref": "#/parameters/body.cursos"
          },
          {
            "$ref": "#/parameters/select"
          },
          {
            "$ref": "#/parameters/preferPost"
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          }
        },
        "tags": [
          "cursos"
        ]
      },
      "delete": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.cursos.id"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.descricaogeral"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.duracao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahoraria"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahorariaestagio"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahorariaatividadescomplementares"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.mediarequerida"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.frequenciarequerida"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.nivelensino"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.grauconferido"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.tituloconferido"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.valorinscricao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.valormensalidade"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.layoutnotas"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.idcenso"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.bibliotecavirtual"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.portariarecredenciamento"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.exibircrm"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.exibirbibliotecavirtual"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.situacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.numero_id"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "cursos"
        ]
      },
      "patch": {
        "parameters": [
          {
            "$ref": "#/parameters/rowFilter.cursos.id"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.nome"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.descricaogeral"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.duracao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahoraria"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahorariaestagio"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.cargahorariaatividadescomplementares"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.mediarequerida"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.frequenciarequerida"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.nivelensino"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.grauconferido"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.tituloconferido"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.valorinscricao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.valormensalidade"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.layoutnotas"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.idcenso"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.bibliotecavirtual"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.portariarecredenciamento"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.exibircrm"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.exibirbibliotecavirtual"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.situacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.datacriacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.dataatualizacao"
          },
          {
            "$ref": "#/parameters/rowFilter.cursos.numero_id"
          },
          {
            "$ref": "#/parameters/body.cursos"
          },
          {
            "$ref": "#/parameters/preferReturn"
          }
        ],
        "responses": {
          "204": {
            "description": "No Content"
          }
        },
        "tags": [
          "cursos"
        ]
      }
    },
...
  },
  "definitions": {
...
    "cursos": {
      "required": [
        "id",
        "nome",
        "numero_id"
      ],
      "properties": {
        "id": {
          "description": "Note:\nThis is a Primary Key.<pk/>",
          "format": "integer",
          "type": "integer"
        },
        "nome": {
          "format": "character varying",
          "maxLength": 255,
          "type": "string"
        },
        "descricaogeral": {
          "format": "text",
          "type": "string"
        },
        "duracao": {
          "format": "integer",
          "type": "integer"
        },
        "cargahoraria": {
          "format": "integer",
          "type": "integer"
        },
        "cargahorariaestagio": {
          "default": 0,
          "format": "integer",
          "type": "integer"
        },
        "cargahorariaatividadescomplementares": {
          "default": 0,
          "format": "integer",
          "type": "integer"
        },
        "mediarequerida": {
          "default": 6,
          "format": "numeric",
          "type": "number"
        },
        "frequenciarequerida": {
          "default": 75,
          "format": "integer",
          "type": "integer"
        },
        "nivelensino": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "grauconferido": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "tituloconferido": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "valorinscricao": {
          "format": "numeric",
          "type": "number"
        },
        "valormensalidade": {
          "format": "numeric",
          "type": "number"
        },
        "layoutnotas": {
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "idcenso": {
          "format": "character varying",
          "maxLength": 20,
          "type": "string"
        },
        "bibliotecavirtual": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "portariarecredenciamento": {
          "format": "character varying",
          "maxLength": 100,
          "type": "string"
        },
        "exibircrm": {
          "default": false,
          "format": "boolean",
          "type": "boolean"
        },
        "exibirbibliotecavirtual": {
          "default": false,
          "format": "boolean",
          "type": "boolean"
        },
        "situacao": {
          "default": "ATIVO",
          "format": "character varying",
          "maxLength": 50,
          "type": "string"
        },
        "datacriacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "dataatualizacao": {
          "default": "CURRENT_TIMESTAMP",
          "format": "timestamp without time zone",
          "type": "string"
        },
        "numero_id": {
          "format": "integer",
          "type": "integer"
        }
      },
      "type": "object"
    },
...
  },
  "parameters": {
...
    "preferReturn": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferCount": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "count=none"
      ],
      "in": "header",
      "type": "string"
    },
    "preferPost": {
      "name": "Prefer",
      "description": "Preference",
      "required": false,
      "enum": [
        "return=representation",
        "return=minimal",
        "return=none",
        "resolution=ignore-duplicates",
        "resolution=merge-duplicates"
      ],
      "in": "header",
      "type": "string"
    },
    "select": {
      "name": "select",
      "description": "Filtering Columns",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "order": {
      "name": "order",
      "description": "Ordering",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "range": {
      "name": "Range",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "header",
      "type": "string"
    },
    "rangeUnit": {
      "name": "Range-Unit",
      "description": "Limiting and Pagination",
      "required": false,
      "default": "items",
      "in": "header",
      "type": "string"
    },
    "offset": {
      "name": "offset",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "limit": {
      "name": "limit",
      "description": "Limiting and Pagination",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
    "body.cursos": {
      "name": "cursos",
      "description": "cursos",
      "required": false,
      "in": "body",
      "schema": {
        "$ref": "#/definitions/cursos"
      }
    },
    "rowFilter.cursos.id": {
      "name": "id",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.nome": {
      "name": "nome",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.descricaogeral": {
      "name": "descricaogeral",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.duracao": {
      "name": "duracao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.cargahoraria": {
      "name": "cargahoraria",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.cargahorariaestagio": {
      "name": "cargahorariaestagio",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.cargahorariaatividadescomplementares": {
      "name": "cargahorariaatividadescomplementares",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.mediarequerida": {
      "name": "mediarequerida",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.frequenciarequerida": {
      "name": "frequenciarequerida",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.nivelensino": {
      "name": "nivelensino",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.grauconferido": {
      "name": "grauconferido",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.tituloconferido": {
      "name": "tituloconferido",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.valorinscricao": {
      "name": "valorinscricao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.valormensalidade": {
      "name": "valormensalidade",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.layoutnotas": {
      "name": "layoutnotas",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.idcenso": {
      "name": "idcenso",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.bibliotecavirtual": {
      "name": "bibliotecavirtual",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.portariarecredenciamento": {
      "name": "portariarecredenciamento",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.exibircrm": {
      "name": "exibircrm",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.exibirbibliotecavirtual": {
      "name": "exibirbibliotecavirtual",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.situacao": {
      "name": "situacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.datacriacao": {
      "name": "datacriacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.dataatualizacao": {
      "name": "dataatualizacao",
      "required": false,
      "in": "query",
      "type": "string"
    },
    "rowFilter.cursos.numero_id": {
      "name": "numero_id",
      "required": false,
      "in": "query",
      "type": "string"
    },
...
  },
...
}